move util methods to util class
