+++ georocket-server/src/main/java/io/georocket/http/StoreEndpoint.java
@@ -1,20 +1,15 @@
 package io.georocket.http;
 
-import static io.georocket.util.ThrowableHelper.throwableToCode;
-import static io.georocket.util.ThrowableHelper.throwableToMessage;
-
+import io.georocket.util.HttpUtils;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.regex.Pattern;
 
-import io.georocket.ServerAPIException;
-import io.vertx.core.eventbus.ReplyException;
 import org.apache.commons.lang3.StringEscapeUtils;
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.http.ParseException;
@@ -52,7 +47,6 @@ import io.vertx.ext.web.RoutingContext;
 import io.vertx.rx.java.ObservableFuture;
 import io.vertx.rx.java.RxHelper;
 import rx.Observable;
-import rx.Single;
 
 /**
  * An HTTP endpoint handling requests related to the GeoRocket data store
@@ -209,7 +203,7 @@ public class StoreEndpoint implements Endpoint {
         if (!(err instanceof FileNotFoundException)) {
           log.error("Could not perform query", err);
         }
-        fail(response, err);
+        HttpUtils.fail(response, err);
       });
   }
   
@@ -374,7 +368,7 @@ public class StoreEndpoint implements Endpoint {
         // run importer
         vertx.eventBus().send(AddressConstants.IMPORTER_IMPORT, msg);
       }, err -> {
-        fail(request.response(), err);
+        HttpUtils.fail(request.response(), err);
         err.printStackTrace();
         fs.delete(filepath, ar -> {});
       });
@@ -398,7 +392,7 @@ public class StoreEndpoint implements Endpoint {
           .end();
       }, err -> {
         log.error("Could not delete chunks", err);
-        fail(response, err);
+        HttpUtils.fail(response, err);
       });
   }
 
@@ -413,69 +407,17 @@ public class StoreEndpoint implements Endpoint {
     HttpServerRequest request = context.request();
     String search = request.getParam("search");
 
-    bodyAsJsonObject(request).subscribe(body -> {
+    HttpUtils.bodyAsJsonObject(request).subscribe(body -> {
       body.put("path", path);
       body.put("search", search);
       vertx.eventBus().<JsonObject>send(AddressConstants.INDEXER_UPDATE, body, msg -> {
         if (msg.succeeded()) {
           response.setStatusCode(204).end();
         } else {
-          fail(response, msg.cause());
+          HttpUtils.fail(response, msg.cause());
         }
       });
-    }, err -> fail(response, err));
-  }
-  
-  /**
-   * Let the request fail by setting the correct http error code and an error
-   * description in the body
-   * @param response the response object
-   * @param throwable the cause of the error
-   */
-  private static void fail(HttpServerResponse response, Throwable throwable) {
-    response
-      .setStatusCode(throwableToCode(throwable))
-      .end(errorResponse(throwable));
+    }, err -> HttpUtils.fail(response, err));
   }
 
-  /**
-   * Generate the json error response for a failed request
-   * @param throwable the cause of the error
-   * @return the json string
-   */
-  private static String errorResponse(Throwable throwable) {
-    String msg = throwableToMessage(throwable, "");
-
-    try {
-      return new JsonObject(msg).toString();
-    } catch (Exception e) {
-      if (throwable instanceof ReplyException) {
-        return ServerAPIException.toJson(ServerAPIException.GENERIC_ERROR, msg)
-          .toString();
-      }
-
-      if (throwable instanceof HttpException) {
-        return ServerAPIException.toJson(ServerAPIException.HTTP_ERROR, msg)
-          .toString();
-      }
-
-      return ServerAPIException.toJson(ServerAPIException.GENERIC_ERROR, msg)
-        .toString();
-    }
-  }
-
-  /**
-   * Get request body as JSON object.
-   * @param request the request to extract to body from
-   * @return Single holding the JSON as soon as the request is processed
-   */
-  private static Single<JsonObject> bodyAsJsonObject(HttpServerRequest request) {
-    return Single.create(singleSubscriber -> {
-      Buffer buffer = Buffer.buffer();
-      request.handler(buffer::appendBuffer);
-      request.exceptionHandler(singleSubscriber::onError);
-      request.endHandler(v -> singleSubscriber.onSuccess(
-          new JsonObject(buffer.toString(StandardCharsets.UTF_8))));
-    });
-  }
 }
